Really the title says it all
This is just a ridiculously simple time tracker for my personal use written in Python talking to SQLite. I put the database file in an owncloud folder to sync it between my machines, again to keep it stupidly simple. Really, this is just a starting point for a solution to a problem I choose to solve for myself.